+++ education-back/src/main/java/com/javagpt/back/entity/Comment.java
@@ -3,6 +3,7 @@ package com.javagpt.back.entity;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
+
 import java.time.LocalDateTime;
 import java.io.Serializable;
 
@@ -11,10 +12,6 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 /**
- * <p>
- * 
- * </p>
- *
  * @author zqy
  * @since 2023-10-17
  */
@@ -24,8 +21,8 @@ public class Comment implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-      @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    @TableId(value = "id", type = IdType.AUTO)
+    private Long id;
 
     /**
      * 评论内容
@@ -52,5 +49,4 @@ public class Comment implements Serializable {
 
     @TableField(exist = false)
     private UserPO user;
-
 }
